requirements.txt
PyYAML==6.0.2
hnswlib==0.8.0
hydra-core==1.3.2
omegaconf==2.3.0
pandas==2.2.3
peft==0.13.2
pytest==8.3.3
scikit-learn==1.5.2
sentence-transformers==3.4.0
torch-optimi==0.2.1
torchao==0.6.1
torchcontrib==0.0.2
tqdm==4.67.0
transformers==4.46.2
wandb